Update README.md
Every sports event generates a lot of data which we can use to analyze the performance of players, teams, and many highlights of the game. As the ICC Men’s Odi world cup has just finished, it has generated a lot of data we can use to summarize the event. So, if you want to learn how to analyze a sports event like the Odi world cup, this article is for you. This article will take you through the task of Odi World Cup 2023 analysis using Python.